changed result to 32bit uint

+++ adwandler.c
@@ -121,9 +121,9 @@ double calc_he_level(double res_x, double res_min, double res_max)
 
 
 ///reads the given channel and returns converted ADc-value (average of "avr" measurements)
-uint16_t readChannel_ILM(uint8_t channel, uint8_t avg)
+uint32_t readChannel_ILM(uint8_t channel, uint16_t avg)
 {
-	uint16_t result = 0;
+	uint32_t result = 0;
 	
 	ADMUX = channel;						// Kanal waehlen
 	ADMUX |= (1<<REFS1) | (1<<REFS0); 	// 2,56
@@ -131,7 +131,7 @@ uint16_t readChannel_ILM(uint8_t channel, uint8_t avg)
 	//ADCSRA = (1<<ADEN) | (1<<ADPS2) | (1<<ADPS1) | (1<<ADPS0); //strom an
 	
 	//Messung - Mittelwert aus "avg" aufeinanderfolgenden Wandlungen
-	for( uint8_t i=0; i<avg; i++ )
+	for( uint16_t i=0; i<avg; i++ )
 	{
 		ADCSRA |= (1<<ADSC);            	// eine Wandlung "single conversion"
 		while ( ADCSRA & (1<<ADSC) );		// auf Abschluss der Konvertierung warten
@@ -141,7 +141,7 @@ uint16_t readChannel_ILM(uint8_t channel, uint8_t avg)
 	
 	//ADCSRA &= ~(1<<ADEN); // ADC deaktivieren
 
-	return result/avg;
+	return result/((uint32_t)avg);
 }
